+++ app/workers/recurring_schedule_worker.rb
@@ -1,0 +1,7 @@
+class RecurringScheduleWorker
+  include Sidekiq::Worker
+
+  def perform
+    RedmineRecurringTasks::IssueChecker.new(Setting.plugin_redmine_recurring_tasks).call
+  end
+end